# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: PrepVI R1 API
  version: 1.0.0
  description: Same-origin API contract for PrepVI US-01–30. Sensitive fields are intentionally absent from error schemas.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /auth/register:
    post:
      security: []
      operationId: register
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RegisterInput" } } } }
      responses: { "201": { $ref: "#/components/responses/User" }, "409": { $ref: "#/components/responses/Error" } }
  /auth/verify-email:
    post:
      security: []
      operationId: verifyEmail
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/TokenInput" } } } }
      responses: { "200": { $ref: "#/components/responses/User" }, "400": { $ref: "#/components/responses/Error" } }
  /auth/login:
    post:
      security: []
      operationId: login
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/LoginInput" } } } }
      responses: { "200": { $ref: "#/components/responses/Login" }, "401": { $ref: "#/components/responses/Error" } }
  /auth/logout:
    post:
      operationId: logout
      responses: { "204": { description: Session revoked }, "403": { $ref: "#/components/responses/Error" } }
...
  /auth/forgot-password:
    post:
      security: []
      operationId: forgotPassword
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "202": { $ref: "#/components/responses/Success" } }
  /auth/reset-password:
    post:
      security: []
      operationId: resetPassword
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "204": { description: Password changed }, "400": { $ref: "#/components/responses/Error" } }
  /auth/accept-admin-invite:
    post:
      security: []
      operationId: acceptAdminInvite
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/User" }, "400": { $ref: "#/components/responses/Error" } }
  /me:
    get:
      operationId: me
      responses: { "200": { $ref: "#/components/responses/User" }, "401": { $ref: "#/components/responses/Error" } }
  /student-profile:
    get:
      operationId: getStudentProfile
      responses: { "200": { $ref: "#/components/responses/Success" } }
    patch:
      operationId: updateStudentProfile
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
...
  /bookings:
    get:
      operationId: listBookings
      parameters:
        - { name: state, in: query, schema: { $ref: "#/components/schemas/BookingStatus" } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { "200": { $ref: "#/components/responses/Page" } }
    post:
      operationId: createBooking
      parameters: [ { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Booking" }, "409": { $ref: "#/components/responses/Error" } }
  /bookings/{bookingId}:
    parameters: [ { $ref: "#/components/parameters/BookingId" } ]
    get:
      operationId: getBooking
      responses: { "200": { $ref: "#/components/responses/Booking" }, "404": { $ref: "#/components/responses/Error" } }
  /bookings/{bookingId}/transitions:
    parameters: [ { $ref: "#/components/parameters/BookingId" }, { $ref: "#/components/parameters/IdempotencyKey" } ]
    post:
      operationId: transitionBooking
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Booking" }, "202": { $ref: "#/components/responses/Booking" }, "409": { $ref: "#/components/responses/Error" } }
  /bookings/{bookingId}/meeting-link:
    parameters: [ { $ref: "#/components/parameters/BookingId" } ]
    put:
      operationId: saveMeetingLink
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /bookings/{bookingId}/meeting-link-failures:
    parameters: [ { $ref: "#/components/parameters/BookingId" } ]
    post:
      operationId: reportMeetingLinkFailure
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "202": { $ref: "#/components/responses/Success" } }
  /bookings/{bookingId}/feedback:
    parameters: [ { $ref: "#/components/parameters/BookingId" } ]
    get:
      operationId: getFeedback
      responses: { "200": { $ref: "#/components/responses/Success" }, "404": { $ref: "#/components/responses/Error" } }
    post:
      operationId: createFeedback
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /bookings/{bookingId}/feedback/apply:
    parameters: [ { $ref: "#/components/parameters/BookingId" } ]
    post:
      operationId: applyFeedbackToPlan
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /bookings/{bookingId}/completion-disputes:
    parameters: [ { $ref: "#/components/parameters/BookingId" } ]
    post:
      operationId: createCompletionDispute
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "202": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /bookings/{bookingId}/operation-cases/{caseId}/actions:
    parameters:
      - { $ref: "#/components/parameters/BookingId" }
      - { $ref: "#/components/parameters/CaseId" }
      - { $ref: "#/components/parameters/IdempotencyKey" }
    post:
      operationId: resolveParticipantOperationCase
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /bookings/{bookingId}/review:
    parameters: [ { $ref: "#/components/parameters/BookingId" } ]
    post:
      operationId: createReview
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: __Host-prepvi_session }
  parameters:
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 200 } }
...
    BookingId: { name: bookingId, in: path, required: true, schema: { type: string, format: uuid } }
    CaseId: { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
  requestBodies:
    JsonInput:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  schemas:
    Recovery:
      type: object
      required: [kind, retryable, retryAfterSeconds]
      properties:
        kind: { type: string, enum: [RETRY_SAFE, EDIT_INPUT, REUPLOAD, PASTE_TEXT, SELECT_ANOTHER_SLOT, WAIT, CONTACT_SUPPORT, NONE] }
        retryable: { type: boolean }
        retryAfterSeconds: { type: [integer, "null"] }
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, fieldErrors, recovery]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string, format: uuid }
        fieldErrors: { type: object, additionalProperties: { type: string } }
        recovery: { $ref: "#/components/schemas/Recovery" }
    User:
      type: object
      required: [id, email, displayName, roles]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        status: { type: string }
        roles: { type: array, items: { type: string, enum: [STUDENT, MENTOR, ADMIN] } }
    RegisterInput:
      type: object
      required: [email, password, displayName]
      properties: { email: { type: string, format: email }, password: { type: string, minLength: 8, writeOnly: true }, displayName: { type: string, minLength: 2 } }
    LoginInput:
      type: object
      required: [email, password]
      properties: { email: { type: string, format: email }, password: { type: string, writeOnly: true } }
    TokenInput:
      type: object
      required: [token]
      properties: { token: { type: string, writeOnly: true } }
...
    BookingStatus: { type: string, enum: [PENDING, CONFIRMED, RESCHEDULE_PROPOSED, REJECTED, CANCELLED, COMPLETED, NO_SHOW] }
...
    Booking:
      type: object
      required: [id, mentorId, slotId, status, startsAt, endsAt, version]
      properties:
        id: { type: string, format: uuid }
        mentorId: { type: string, format: uuid }
        slotId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/BookingStatus" }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        version: { type: integer }
    PageInfo:
      type: object
      required: [page, pageSize, total]
      properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer } }
  responses:
    Success: { description: Successful resource response, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
    Page: { description: Paginated response, content: { application/json: { schema: { type: object, required: [items, pageInfo], properties: { items: { type: array, items: { type: object, additionalProperties: true } }, pageInfo: { $ref: "#/components/schemas/PageInfo" } } } } } }
    Error: { description: Safe actionable error, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    User: { description: Safe user projection, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
    Login: { description: Authenticated session, content: { application/json: { schema: { type: object, required: [user, csrfToken], properties: { user: { $ref: "#/components/schemas/User" }, csrfToken: { type: string } } } } } }
...
    Booking: { description: Booking projection, content: { application/json: { schema: { $ref: "#/components/schemas/Booking" } } } }